file_io: remove package
This collapses the _file_io and _path modules into the trio package, and moves
_helpers to _util
