+++ trio/__init__.py
@@ -53,6 +53,9 @@ __all__ += _network.__all__
 from ._file_io import *
 __all__ += _file_io.__all__
 
+from ._path import *
+__all__ += _path.__all__
+
 # Imported by default
 from . import socket
 from . import abc
