--- trio/_file_io/__init__.py
@@ -1,7 +1,0 @@
-__all__ = []
-
-from ._file_io import *
-__all__ += _file_io.__all__
-
-from ._path import *
-__all__ += _path.__all__
