+++ trio/_util.py
@@ -188,3 +188,22 @@ class UnLock:
 
     async def __aexit__(self, *args):
         return self.sync.__exit__()
+
+
+def async_wraps(cls, wrapped_cls, attr_name):
+    """Similar to wraps, but for async wrappers of non-async functions.
+
+    """
+    def decorator(func):
+        func.__name__ = attr_name
+        func.__qualname__ = '.'.join((cls.__qualname__,
+                                      attr_name))
+
+        func.__doc__ = """Like :meth:`~{}.{}.{}`, but async.
+
+        """.format(wrapped_cls.__module__,
+                   wrapped_cls.__qualname__,
+                   attr_name)
+
+        return func
+    return decorator
